+++ nlp/stemmer/id/dictionary.php
