+++ nlp/stopword/lang/en.php
