+++ nlp/transformer/lsa.php
@@ -22,15 +22,15 @@ class lsa implements transformer {
    * Perform latent semantic analysis to get the most important topic of the word with dimensional reduction
    */
   public function transform() {
-    $transformed = (new svd($this->matrix))->truncate()->transform();
-    $original = $this->matrix->original();
+    $svd = (new svd($this->matrix))->transform();
+    $transformed = [];
 
-    foreach ($original as $i => $_) {
-      $original[$i] = array_combine(
-        array_keys($original[0]), array_values($transformed[$i])
+    foreach ($this->matrix->original() as $i => $_) {
+      $transformed[$i] = array_combine(
+        array_keys($this->matrix->original()[0]), array_values($svd[$i])
       );
     }
 
-    return $original;
+    return $transformed;
   }
 }
